+++ packages/accounts-base/accounts_common.js
@@ -89,6 +89,9 @@ export class AccountsCommon {
   // - ambiguousErrorMessages {Boolean}
   //     Return ambiguous error messages from login failures to prevent
   //     user enumeration.
+  // - bcryptRounds {Number}
+  //     Allows override of number of bcrypt rounds (aka work factor) used
+  //     to store passwords.
 
   /**
    * @summary Set global accounts options.
@@ -135,7 +138,7 @@ export class AccountsCommon {
     // validate option keys
     var VALID_KEYS = ["sendVerificationEmail", "forbidClientAccountCreation", "passwordEnrollTokenExpirationInDays",
                       "restrictCreationByEmailDomain", "loginExpirationInDays", "passwordResetTokenExpirationInDays",
-                      "ambiguousErrorMessages"];
+                      "ambiguousErrorMessages", "bcryptRounds"];
     _.each(_.keys(options), function (key) {
       if (!_.contains(VALID_KEYS, key)) {
         throw new Error("Accounts.config: Invalid key: " + key);